requirements.txt
kucoin-python
pandas
matplotlib
python-dotenv
setuptools<81
pytest
notebook
